chore: onClickCopy, Delete 위치 수정
